Caching the inverse of a matrix
This assignment consists of two functions named cacheSolve and makeCacheMatrix.
The makeCacheMatrix Function
This function creates a special "matrix" object whose inverse can be cached. Essentially, it stores a matrix and provides a place to store its inverse. It also includes functional methods such as setting the matrix, getting the matrix, setting the inverse of the matrix, and getting the inverse of the matrix.
cacheSolve Function
This function calculates the inverse of the special "matrix" object returned by makeCacheMatrix. If the inverse is already cached, it retrieves it directly from the cache; otherwise, it calculates it and stores it in the cache.